+++ src/dev/theskidster/xjge/audio/Source.java
@@ -1,7 +1,7 @@
 package dev.theskidster.xjge.audio;
 
 import static org.lwjgl.openal.AL11.*;
-import dev.theskidster.xjge.util.ErrorUtil;
+import dev.theskidster.xjge.main.ErrorUtil;
 import org.joml.Vector3f;
 
 /**