+++ src/dev/theskidster/xjge/graphics/Cubemap.java
@@ -1,7 +1,7 @@
 package dev.theskidster.xjge.graphics;
 
 import dev.theskidster.xjge.main.App;
-import dev.theskidster.xjge.util.ErrorUtil;
+import dev.theskidster.xjge.main.ErrorUtil;
 import dev.theskidster.xjge.main.Logger;
 import java.io.IOException;
 import java.io.InputStream;