+++ src/dev/theskidster/xjge/main/Viewport.java
@@ -13,7 +13,6 @@ import dev.theskidster.xjge.graphics.Graphics;
 import dev.theskidster.xjge.puppet.freecam.Freecam;
 import dev.theskidster.xjge.shader.core.ShaderCore;
 import dev.theskidster.xjge.ui.Component;
-import dev.theskidster.xjge.util.ErrorUtil;
 
 /**
  * @author J Hoffman