+++ src/dev/theskidster/xjge/ui/Icon.java
@@ -5,7 +5,7 @@ import dev.theskidster.xjge.graphics.Graphics;
 import dev.theskidster.xjge.graphics.SpriteSheet;
 import dev.theskidster.xjge.graphics.Texture;
 import dev.theskidster.xjge.shader.core.ShaderCore;
-import dev.theskidster.xjge.util.ErrorUtil;
+import dev.theskidster.xjge.main.ErrorUtil;
 import dev.theskidster.xjge.main.Logger;
 import java.util.HashMap;
 import java.util.Map;