+++ src/dev/theskidster/xjge/ui/Polygon.java
@@ -3,7 +3,7 @@ package dev.theskidster.xjge.ui;
 import dev.theskidster.xjge.graphics.Graphics;
 import dev.theskidster.xjge.shader.core.ShaderCore;
 import dev.theskidster.xjge.util.Color;
-import dev.theskidster.xjge.util.ErrorUtil;
+import dev.theskidster.xjge.main.ErrorUtil;
 import org.joml.Vector2f;
 import org.joml.Vector3f;
 import static org.lwjgl.opengl.GL30.*;