+++ src/dev/theskidster/xjge/ui/RectangleBatch.java
@@ -6,7 +6,7 @@ import org.lwjgl.system.MemoryUtil;
 import dev.theskidster.xjge.graphics.Graphics;
 import dev.theskidster.xjge.shader.core.ShaderCore;
 import dev.theskidster.xjge.util.Color;
-import dev.theskidster.xjge.util.ErrorUtil;
+import dev.theskidster.xjge.main.ErrorUtil;
 
 /**
  * @author J Hoffman